Rename `/servers` page to `/browse`
